fix(backtracking): correct stock selling strategy calculation
- Fix type casting issue in profit calculation formula
- Add proper long type conversion for price values
- Correct arithmetic operations to prevent integer overflow
- Ensure accurate summation of trading strategy results

+++ src/backtracking/BestTimeToBuyAndSellStockUsingStrategy.java
@@ -30,7 +30,7 @@ public class BestTimeToBuyAndSellStockUsingStrategy {
             long s = strategy[i];
             base += p * s;
             sum += p * (1L - s)
-                    - (long) prices[i - m]
+                    - prices[i - m]
                     + (long) prices[i - k] * (long) strategy[i - k];
             if (sum > best) best = sum;
         }
